+++ analysis/src/CompletionBackEnd.ml
@@ -980,14 +980,75 @@ let findLocalCompletionsWithOpens ~pos ~(env : QueryEnv.t) ~prefix ~exact ~opens
     (* There's no local completion for fields *)
     []
 
+let instantiateType ~typeParams ~typeArgs (t : Types.type_expr) =
+  if typeParams = [] || typeArgs = [] then t
+  else
+    let rec applySub tp ta t =
+      match (tp, ta) with
+      | t1 :: tRest1, t2 :: tRest2 ->
+        if t1 = t then t2 else applySub tRest1 tRest2 t
+      | [], _ | _, [] -> t
+    in
+    let rec loop (t : Types.type_expr) =
+      match t.desc with
+      | Tlink t -> loop t
+      | Tvar _ -> applySub typeParams typeArgs t
+      | Tunivar _ -> t
+      | Tconstr (path, args, memo) ->
+        {t with desc = Tconstr (path, args |> List.map loop, memo)}
+      | Tsubst t -> loop t
+      | Tvariant rd -> {t with desc = Tvariant (rowDesc rd)}
+      | Tnil -> t
+      | Tarrow (lbl, t1, t2, c) ->
+        {t with desc = Tarrow (lbl, loop t1, loop t2, c)}
+      | Ttuple tl -> {t with desc = Ttuple (tl |> List.map loop)}
+      | Tobject (t, r) -> {t with desc = Tobject (loop t, r)}
+      | Tfield (n, k, t1, t2) -> {t with desc = Tfield (n, k, loop t1, loop t2)}
+      | Tpoly (t, []) -> loop t
+      | Tpoly (t, tl) -> {t with desc = Tpoly (loop t, tl |> List.map loop)}
+      | Tpackage (p, l, tl) ->
+        {t with desc = Tpackage (p, l, tl |> List.map loop)}
+    and rowDesc (rd : Types.row_desc) =
+      let row_fields =
+        rd.row_fields |> List.map (fun (l, rf) -> (l, rowField rf))
+      in
+      let row_more = loop rd.row_more in
+      let row_name =
+        match rd.row_name with
+        | None -> None
+        | Some (p, tl) -> Some (p, tl |> List.map loop)
+      in
+      {rd with row_fields; row_more; row_name}
+    and rowField (rf : Types.row_field) =
+      match rf with
+      | Rpresent None -> rf
+      | Rpresent (Some t) -> Rpresent (Some (loop t))
+      | Reither (b1, tl, b2, r) -> Reither (b1, tl |> List.map loop, b2, r)
+      | Rabsent -> Rabsent
+    in
+    loop t
+
 let rec extractRecordType ~env ~package (t : Types.type_expr) =
   match t.desc with
   | Tlink t1 | Tsubst t1 | Tpoly (t1, []) -> extractRecordType ~env ~package t1
-  | Tconstr (path, _, _) -> (
+  | Tconstr (path, typeArgs, _) -> (
     match References.digConstructor ~env ~package path with
     | Some (env, ({item = {kind = Record fields}} as typ)) ->
+      let typeParams = typ.item.decl.type_params in
+      let fields =
+        fields
+        |> List.map (fun field ->
+               let fieldTyp =
+                 field.typ |> instantiateType ~typeParams ~typeArgs
+               in
+               {field with typ = fieldTyp})
+      in
       Some (env, fields, typ)
-    | Some (env, {item = {decl = {type_manifest = Some t1}}}) ->
+    | Some
+        ( env,
+          {item = {decl = {type_manifest = Some t1; type_params = typeParams}}}
+        ) ->
+      let t1 = t1 |> instantiateType ~typeParams ~typeArgs in
       extractRecordType ~env ~package t1
     | _ -> None)
   | _ -> None
@@ -996,9 +1057,13 @@ let rec extractObjectType ~env ~package (t : Types.type_expr) =
   match t.desc with
   | Tlink t1 | Tsubst t1 | Tpoly (t1, []) -> extractObjectType ~env ~package t1
   | Tobject (tObj, _) -> Some (env, tObj)
-  | Tconstr (path, _, _) -> (
+  | Tconstr (path, typeArgs, _) -> (
     match References.digConstructor ~env ~package path with
-    | Some (env, {item = {decl = {type_manifest = Some t1}}}) ->
+    | Some
+        ( env,
+          {item = {decl = {type_manifest = Some t1; type_params = typeParams}}}
+        ) ->
+      let t1 = t1 |> instantiateType ~typeParams ~typeArgs in
       extractObjectType ~env ~package t1
     | _ -> None)
   | _ -> None
@@ -1008,9 +1073,14 @@ let extractFunctionType ~env ~package typ =
     match t.desc with
     | Tlink t1 | Tsubst t1 | Tpoly (t1, []) -> loop ~env acc t1
     | Tarrow (label, tArg, tRet, _) -> loop ~env ((label, tArg) :: acc) tRet
-    | Tconstr (path, _, _) -> (
+    | Tconstr (path, typeArgs, _) -> (
       match References.digConstructor ~env ~package path with
-      | Some (env, {item = {decl = {type_manifest = Some t1}}}) ->
+      | Some
+          ( env,
+            {
+              item = {decl = {type_manifest = Some t1; type_params = typeParams}};
+            } ) ->
+        let t1 = t1 |> instantiateType ~typeParams ~typeArgs in
         loop ~env acc t1
       | _ -> (List.rev acc, t))
     | _ -> (List.rev acc, t)
@@ -1234,17 +1304,17 @@ let rec getCompletionsForContextPath ~package ~opens ~rawOpens ~allFiles ~pos
           | _ :: rest -> List.rev rest
           | [] -> [])
       in
-      let getConstr typ =
+      let getConstrPath typ =
         match typ.Types.desc with
-        | Tconstr (path, _, _)
-        | Tlink {desc = Tconstr (path, _, _)}
-        | Tsubst {desc = Tconstr (path, _, _)}
-        | Tpoly ({desc = Tconstr (path, _, _)}, []) ->
+        | Tconstr (path, _typeArgs, _)
+        | Tlink {desc = Tconstr (path, _typeArgs, _)}
+        | Tsubst {desc = Tconstr (path, _typeArgs, _)}
+        | Tpoly ({desc = Tconstr (path, _typeArgs, _)}, []) ->
           Some path
         | _ -> None
       in
       let fromType typ =
-        match getConstr typ with
+        match getConstrPath typ with
         | None -> None
         | Some path -> Some (getModulePath path)
       in
@@ -1357,14 +1427,37 @@ let processCompletable ~debug ~package ~scope ~env ~pos ~forHover
   | Cjsx (componentPath, prefix, identsSeen) ->
     let labels =
       match componentPath @ ["make"] |> findTypeOfValue with
-      | Some (typ, _env) ->
-        let rec getFields (texp : Types.type_expr) =
+      | Some (typ, make_env) ->
+        let rec getFieldsV3 (texp : Types.type_expr) =
           match texp.desc with
           | Tfield (name, _, t1, t2) ->
-            let fields = t2 |> getFields in
+            let fields = t2 |> getFieldsV3 in
             if name = "children" then fields else (name, t1) :: fields
-          | Tlink te | Tsubst te | Tpoly (te, []) -> te |> getFields
+          | Tlink te | Tsubst te | Tpoly (te, []) -> te |> getFieldsV3
           | Tvar None -> []
+          | _ -> []
+        in
+        let getFieldsV4 ~path ~typeArgs =
+          match References.digConstructor ~env:make_env ~package path with
+          | Some
+              ( _env,
+                {
+                  item =
+                    {
+                      decl =
+                        {
+                          type_kind = Type_record (labelDecls, _repr);
+                          type_params = typeParams;
+                        };
+                    };
+                } ) ->
+            labelDecls
+            |> List.map (fun (ld : Types.label_declaration) ->
+                   let name = Ident.name ld.ld_id in
+                   let t =
+                     ld.ld_type |> instantiateType ~typeParams ~typeArgs
+                   in
+                   (name, t))
           | _ -> []
         in
         let rec getLabels (t : Types.type_expr) =
@@ -1379,9 +1472,14 @@ let processCompletable ~debug ~package ~scope ~env ~pos ~forHover
                 },
                 _,
                 _ ) ->
-            getFields tObj
+            (* JSX V3 *)
+            getFieldsV3 tObj
+          | Tarrow (Nolabel, {desc = Tconstr (path, typeArgs, _)}, _, _)
+            when Path.last path = "props" ->
+            (* JSX V4 *)
+            getFieldsV4 ~path ~typeArgs
           | Tconstr
-              ( path,
+              ( clPath,
                 [
                   {
                     desc =
@@ -1391,8 +1489,14 @@ let processCompletable ~debug ~package ~scope ~env ~pos ~forHover
                   _;
                 ],
                 _ )
-            when Path.name path = "React.componentLike" ->
-            getFields tObj
+            when Path.name clPath = "React.componentLike" ->
+            (* JSX V3 external or interface *)
+            getFieldsV3 tObj
+          | Tconstr (clPath, [{desc = Tconstr (path, typeArgs, _)}; _], _)
+            when Path.name clPath = "React.componentLike"
+                 && Path.last path = "props" ->
+            (* JSX V4 external or interface *)
+            getFieldsV4 ~path ~typeArgs
           | _ -> []
         in
         typ |> getLabels
@@ -1410,6 +1514,7 @@ let processCompletable ~debug ~package ~scope ~env ~pos ~forHover
       (labels
       |> List.filter (fun (name, _t) ->
              Utils.startsWith name prefix
+             && name <> "key"
              && (forHover || not (List.mem name identsSeen)))
       |> List.map mkLabel)
       @ keyLabels
@@ -1668,9 +1773,18 @@ Note: The `@react.component` decorator requires the react-jsx config to be set i
           | Tarrow ((Labelled l | Optional l), tArg, tRet, _) ->
             (l, tArg) :: getLabels ~env tRet
           | Tarrow (Nolabel, _, tRet, _) -> getLabels ~env tRet
-          | Tconstr (path, _, _) -> (
+          | Tconstr (path, typeArgs, _) -> (
             match References.digConstructor ~env ~package path with
-            | Some (env, {item = {decl = {type_manifest = Some t1}}}) ->
+            | Some
+                ( env,
+                  {
+                    item =
+                      {
+                        decl =
+                          {type_manifest = Some t1; type_params = typeParams};
+                      };
+                  } ) ->
+              let t1 = t1 |> instantiateType ~typeParams ~typeArgs in
               getLabels ~env t1
             | _ -> [])
           | _ -> []