+++ analysis/src/Hover.ml
@@ -2,6 +2,50 @@ open SharedTypes
 
 let codeBlock code = Printf.sprintf "```rescript\n%s\n```" code
 
+(* Light weight, hopefully-enough-for-the-purpose fn to encode URI components.
+   Built to handle the reserved characters listed in
+   https://en.wikipedia.org/wiki/Percent-encoding. Note that this function is not
+   general purpose, rather it's currently only for URL encoding the argument list
+   passed to command links in markdown. *)
+let encodeURIComponent text =
+  let ln = String.length text in
+  let buf = Buffer.create ln in
+  let rec loop i =
+    if i < ln then (
+      (match text.[i] with
+      | '"' -> Buffer.add_string buf "%22"
+      | '\'' -> Buffer.add_string buf "%22"
+      | ':' -> Buffer.add_string buf "%3A"
+      | ';' -> Buffer.add_string buf "%3B"
+      | '/' -> Buffer.add_string buf "%2F"
+      | '\\' -> Buffer.add_string buf "%5C"
+      | ',' -> Buffer.add_string buf "%2C"
+      | '&' -> Buffer.add_string buf "%26"
+      | '[' -> Buffer.add_string buf "%5B"
+      | ']' -> Buffer.add_string buf "%5D"
+      | '#' -> Buffer.add_string buf "%23"
+      | '$' -> Buffer.add_string buf "%24"
+      | '+' -> Buffer.add_string buf "%2B"
+      | '=' -> Buffer.add_string buf "%3D"
+      | '?' -> Buffer.add_string buf "%3F"
+      | '@' -> Buffer.add_string buf "%40"
+      | '%' -> Buffer.add_string buf "%25"
+      | c -> Buffer.add_char buf c);
+      loop (i + 1))
+  in
+  loop 0;
+  Buffer.contents buf
+
+type link = {startPos: Protocol.position; file: string; label: string}
+
+let linkToCommandArgs link =
+  Printf.sprintf "[\"%s\",%i,%i]" link.file link.startPos.line
+    link.startPos.character
+
+let makeGotoCommand link =
+  Printf.sprintf "[%s](command:rescript-vscode.go_to_location?%s)" link.label
+    (encodeURIComponent (linkToCommandArgs link))
+
 let showModuleTopLevel ~docstring ~name (topLevel : Module.item list) =
   let contents =
     topLevel
@@ -36,7 +80,7 @@ let rec showModule ~docstring ~(file : File.t) ~name
   | Some {item = Ident path} ->
     Some ("Unable to resolve module reference " ^ Path.name path)
 
-let newHover ~full:{file; package} locItem =
+let newHover ~full:{file; package} ~supportsMarkdownLinks locItem =
   match locItem.locType with
   | TypeDefinition (name, decl, _stamp) ->
     let typeDef = Shared.declToString name decl in
@@ -98,25 +142,80 @@ let newHover ~full:{file; package} locItem =
          | Const_int64 _ -> "int64"
          | Const_nativeint _ -> "int"))
   | Typed (_, t, locKind) ->
+    let fromConstructorPath ~env path =
+      match References.digConstructor ~env ~package path with
+      | None -> None
+      | Some (env, {extentLoc; item = {decl}}) ->
+        if Utils.isUncurriedInternal path then None
+        else
+          Some
+            ( decl
+              |> Shared.declToString ~printNameAsIs:true
+                   (SharedTypes.pathIdentToString path),
+              extentLoc,
+              env )
+    in
     let fromType ~docstring typ =
       let typeString = codeBlock (typ |> Shared.typeToString) in
-      let extraTypeInfo =
+      let typeDefinitions =
+        (* Expand definitions of types mentioned in typ.
+           If typ itself is a record or variant, search its body *)
         let env = QueryEnv.fromFile file in
-        match typ |> Shared.digConstructor with
-        | None -> None
-        | Some path -> (
-          match References.digConstructor ~env ~package path with
-          | None -> None
-          | Some (_env, {docstring; name = {txt}; item = {decl}}) ->
-            if Utils.isUncurriedInternal path then None
-            else Some (decl |> Shared.declToString txt, docstring))
+        let envToSearch, typesToSearch =
+          match typ |> Shared.digConstructor with
+          | Some path -> (
+            let labelDeclarationsTypes lds =
+              lds |> List.map (fun (ld : Types.label_declaration) -> ld.ld_type)
+            in
+            match References.digConstructor ~env ~package path with
+            | None -> (env, [typ])
+            | Some (env1, {item = {decl}}) -> (
+              match decl.type_kind with
+              | Type_record (lds, _) ->
+                (env1, typ :: (lds |> labelDeclarationsTypes))
+              | Type_variant cds ->
+                ( env1,
+                  cds
+                  |> List.map (fun (cd : Types.constructor_declaration) ->
+                         let fromArgs =
+                           match cd.cd_args with
+                           | Cstr_tuple ts -> ts
+                           | Cstr_record lds -> lds |> labelDeclarationsTypes
+                         in
+                         typ
+                         ::
+                         (match cd.cd_res with
+                         | None -> fromArgs
+                         | Some t -> t :: fromArgs))
+                  |> List.flatten )
+              | _ -> (env, [typ])))
+          | None -> (env, [typ])
+        in
+        let constructors = Shared.findTypeConstructors typesToSearch in
+        constructors
+        |> List.filter_map (fun constructorPath ->
+               match
+                 constructorPath |> fromConstructorPath ~env:envToSearch
+               with
+               | None -> None
+               | Some (typString, extentLoc, env) ->
+                 let startLine, startCol = Pos.ofLexing extentLoc.loc_start in
+                 let linkToTypeDefinitionStr =
+                   if supportsMarkdownLinks then
+                     "\nGo to: "
+                     ^ makeGotoCommand
+                         {
+                           label = "Type definition";
+                           file = Uri.toString env.file.uri;
+                           startPos = {line = startLine; character = startCol};
+                         }
+                   else ""
+                 in
+                 Some
+                   (Shared.markdownSpacing ^ codeBlock typString
+                  ^ linkToTypeDefinitionStr ^ "\n\n---\n"))
       in
-      let typeString, docstring =
-        match extraTypeInfo with
-        | None -> (typeString, docstring)
-        | Some (extra, extraDocstring) ->
-          (typeString ^ "\n\n" ^ codeBlock extra, extraDocstring)
-      in
+      let typeString = typeString :: typeDefinitions |> String.concat "\n\n" in
       (typeString, docstring)
     in
     let parts =